samples/evaluation/deepeval/requirements.txt
deepeval==2.6.8
google-genai==1.9.0

